Centralize indent_tree helper
