+++ batch_program_editor.py
@@ -4,31 +4,8 @@ import logging
 import xml.etree.ElementTree as ET
 import json
 from collections import defaultdict
-from xpm_utils import _parse_xpm_for_rebuild
+from xpm_utils import _parse_xpm_for_rebuild, indent_tree
 from xml.sax.saxutils import escape as xml_escape, unescape as xml_unescape
-
-def indent_tree(tree: ET.ElementTree, space: str = "  ") -> None:
-    """Indent an ElementTree for pretty printing."""
-    if hasattr(ET, "indent"):
-        ET.indent(tree, space=space)
-    else:
-        def _indent(elem: ET.Element, level: int = 0) -> None:
-            i = "\n" + level * space
-            if len(elem):
-                if not elem.text or not elem.text.strip():
-                    elem.text = i + space
-                for child in elem:
-                    _indent(child, level + 1)
-                if not child.tail or not child.tail.strip():  # type: ignore
-                    child.tail = i  # type: ignore
-            if level and (not elem.tail or not elem.tail.strip()):
-                elem.tail = i
-
-        _indent(tree.getroot())
-
-    root = tree.getroot()
-    if not (root.tail and root.tail.endswith("\n")):
-        root.tail = "\n"
 
 from xpm_parameter_editor import (
     set_layer_keytrack,
